+++ typescript/src/index.ts
@@ -9,7 +9,7 @@ export { Network } from "./lib/networks/network";
 export { PartitionID, NetworkContext } from "./lib/networks/network_context";
 
 //Modules
-export { LambdaContext, LambdaID } from "./lib/modules/lambda_context";
+export { ModuleContext, ModuleID } from "./lib/modules/module_context";
 
 //Bulk Command Execution
 export { SchemaVersion } from "./lib/bulk_command_execution/bulk_command_schema_version";
@@ -17,7 +17,7 @@ export { V0BulkCommands, V0SerializableCommand } from "./lib/bulk_command_execut
 export { V0CommandType, V0CommandTypeVisitor } from "./lib/bulk_command_execution/v0_bulk_command_api/v0_command_types";;
 
 //Constructor Calls
-export { newExecCommandArgs, newLoadLambdaArgs, newRegisterFilesArtifactsArgs, newRegisterServiceArgs, newStartServiceArgs, newGetServiceInfoArgs, newRemoveServiceArgs, newPartitionServices, newRepartitionArgs, newPartitionConnections, newWaitForHttpGetEndpointAvailabilityArgs, newWaitForHttpPostEndpointAvailabilityArgs, newExecuteBulkCommandsArgs, newExecuteLambdaArgs, newGetLambdaInfoArgs } from "./lib/constructor_calls";
+export { newExecCommandArgs, newLoadModuleArgs, newRegisterFilesArtifactsArgs, newRegisterServiceArgs, newStartServiceArgs, newGetServiceInfoArgs, newRemoveServiceArgs, newPartitionServices, newRepartitionArgs, newPartitionConnections, newWaitForHttpGetEndpointAvailabilityArgs, newWaitForHttpPostEndpointAvailabilityArgs, newExecuteBulkCommandsArgs, newExecuteModuleArgs, newGetModuleInfoArgs } from "./lib/constructor_calls";
 
 //Kurtosis Core RPC API Bindings
 export { ApiContainerServiceClient } from "./kurtosis_core_rpc_api_bindings/api_container_service_grpc_pb";